[file/#12]:profile 확장자 변경
profile .js라고 되어 있던거 profile.jsx로 변경
